pyproject.toml
[tool.poetry]
name = "tweetlytics"
version = "0.1.4"
description = "Retrieves tweets on a required topic and timeframe using the official twitter API, stores them as a .json and .csv file, performs data cleaning, data analysis and plotting."
authors = ["DSCI_524_2022_Group8"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.3.5"
altair = "^4.2.0"
python-dotenv = "^0.19.2"
requests = "^2.27.1"
arrow = "^1.2.1"
tweepy = "^4.4.0"
textblob = "^0.17.1"
wordcloud = "^1.8.1"
matplotlib = "^3.5.1"
altair-saver = "^0.5.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
codecov = "^2.1.12"
myst-nb = "^0.13.1"
sphinx-autoapi = "^1.8.4"
sphinx-rtd-theme = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.semantic_release]
version_variable = "pyproject.toml:version" # version location
branch = "main"                             # branch to make releases of
changelog_file = "CHANGELOG.md"             # changelog file
build_command = "pip install poetry && poetry build"  # build dists
dist_path = "dist/"                         # where to put dists
upload_to_pypi = false                      # don't auto-upload to PyPI
remove_dist = false                         # don't remove dists
patch_without_tag = true
